split lib3h message handling to own file
