+++ crates/tools/sim_chat/src/lib3h_simchat/mod.rs
@@ -1,0 +1,4 @@
+
+mod handle_chat_event;
+
+pub use handle_chat_event::handle_chat_event;
